ci: push templates to auxilary repos (#1625)
Goes back to before #1314, except the difference now is the template repos have
no activity in them, they're entirely tied to the `template/*` folder
in this repo by this CI job. Testing is still done in this repo.

Actually updating the docs to the new repo URLs is still TODO