fix(paxos): Proposers don't re-commit entries before the latest checkpoint (#1615)
Paxos proposers assign no-ops to holes in the log after receiving a
quorum of p1bs. Acceptors delete log entries before the latest
checkpoint, which means they will show up as holes in the proposer. The
proposer should not propose any operations before the most recent
checkpoint in a quorum of p1bs.